refactor(disconnect): restructure disconnect view [WTEL-9408](https://webitel.atlassian.net/browse/WTEL-9408)
- Replaced the previous disconnect view with a new component structure in DisconnectView.vue.
- Introduced a new SVG icon for the disconnect view.
- Removed unused code and styles from App.vue to streamline the component.
